创建支付单接口做简单幂等处理

+++ pay-service/pay-api/pay-channel-api/src/main/java/com/gloryjie/pay/channel/dto/param/ChargeCreateParam.java
@@ -45,7 +45,7 @@ public class ChargeCreateParam {
      * 支付单金额
      */
     @NotNull
-    private BigDecimal amount;
+    private Long amount;
     /**
      * 购买商品的标题
      */
